% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signalingdata2018}
\alias{signalingdata2018}
\title{Signaling Data 2018}
\format{A data frame with 1636 rows and 116 variables:
\describe{
  \item{\code{StartDate}}{double COLUMN_DESCRIPTION}
  \item{\code{EndDate}}{double COLUMN_DESCRIPTION}
  \item{\code{Status}}{character COLUMN_DESCRIPTION}
  \item{\code{IPAddress}}{character COLUMN_DESCRIPTION}
  \item{\code{Progress}}{integer COLUMN_DESCRIPTION}
  \item{\code{Duration (in seconds)}}{integer COLUMN_DESCRIPTION}
  \item{\code{Finished}}{character COLUMN_DESCRIPTION}
  \item{\code{RecordedDate}}{double COLUMN_DESCRIPTION}
  \item{\code{ResponseId}}{character COLUMN_DESCRIPTION}
  \item{\code{RecipientLastName}}{character COLUMN_DESCRIPTION}
  \item{\code{RecipientFirstName}}{character COLUMN_DESCRIPTION}
  \item{\code{RecipientEmail}}{character COLUMN_DESCRIPTION}
  \item{\code{ExternalReference}}{character COLUMN_DESCRIPTION}
  \item{\code{LocationLatitude}}{double COLUMN_DESCRIPTION}
  \item{\code{LocationLongitude}}{double COLUMN_DESCRIPTION}
  \item{\code{DistributionChannel}}{character COLUMN_DESCRIPTION}
  \item{\code{UserLanguage}}{character COLUMN_DESCRIPTION}
  \item{\code{MO1.1_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.2_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.3_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.4_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.5_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.6_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.7_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.8_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.9_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.10_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.11_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO1.12_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MC1.1_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MC1.2_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{AC1:CC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:CC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC3:CC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC4:CC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC5:CC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC1:CH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:CH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC3:CH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC4:CH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC5:CH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC1:CP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:CP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:CP_1}}{character COLUMN_DESCRIPTION}
  \item{\code{AC3:CP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC4:CP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC1:SC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:SC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC3:SC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC4:SC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC5:SC}}{character COLUMN_DESCRIPTION}
  \item{\code{AC1:SH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:SH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC3:SH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC4:SH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC5:SH}}{character COLUMN_DESCRIPTION}
  \item{\code{AC1:SP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC2:SP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC3:SP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC4:SP}}{character COLUMN_DESCRIPTION}
  \item{\code{AC5:SP}}{character COLUMN_DESCRIPTION}
  \item{\code{MO2.1_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.2_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.3_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.4_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.5_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.6_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.7_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.8_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.9_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.10_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.11_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MO2.12_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MC2.1_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MC2.2_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{MC2.3}}{character COLUMN_DESCRIPTION}
  \item{\code{MC2.4_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{Age}}{integer COLUMN_DESCRIPTION}
  \item{\code{Sex}}{character COLUMN_DESCRIPTION}
  \item{\code{State}}{character COLUMN_DESCRIPTION}
  \item{\code{Children}}{integer COLUMN_DESCRIPTION}
  \item{\code{Siblings}}{character COLUMN_DESCRIPTION}
  \item{\code{BOrder}}{character COLUMN_DESCRIPTION}
  \item{\code{RelStat}}{character COLUMN_DESCRIPTION}
  \item{\code{Ed}}{character COLUMN_DESCRIPTION}
  \item{\code{MTurkID}}{character COLUMN_DESCRIPTION}
  \item{\code{Feedback}}{character COLUMN_DESCRIPTION}
  \item{\code{SC0}}{integer COLUMN_DESCRIPTION}
  \item{\code{Score}}{integer COLUMN_DESCRIPTION}
  \item{\code{MTurkCode}}{integer COLUMN_DESCRIPTION}
  \item{\code{conflict}}{character COLUMN_DESCRIPTION}
  \item{\code{p_info}}{character COLUMN_DESCRIPTION}
  \item{\code{FL_277_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{FL_288_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{FL_298_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{FL_308_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{FL_318_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{FL_207_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{MainOutcome1:Sister,$50,000_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{MainOutcome2:Sister;50_DO}}{character COLUMN_DESCRIPTION}
  \item{\code{pass}}{logical COLUMN_DESCRIPTION}
  \item{\code{wordcount}}{double COLUMN_DESCRIPTION}
  \item{\code{ddupe}}{logical COLUMN_DESCRIPTION}
  \item{\code{latlong}}{character COLUMN_DESCRIPTION}
  \item{\code{dupecount}}{integer COLUMN_DESCRIPTION}
  \item{\code{short}}{logical COLUMN_DESCRIPTION}
  \item{\code{dupe}}{logical COLUMN_DESCRIPTION}
  \item{\code{X1}}{integer COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{countryName}}{character COLUMN_DESCRIPTION}
  \item{\code{asn}}{integer COLUMN_DESCRIPTION}
  \item{\code{isp}}{character COLUMN_DESCRIPTION}
  \item{\code{block}}{integer COLUMN_DESCRIPTION}
  \item{\code{hostname}}{character COLUMN_DESCRIPTION}
  \item{\code{dupe2}}{logical COLUMN_DESCRIPTION}
  \item{\code{signal}}{character COLUMN_DESCRIPTION}
}}
\usage{
signalingdata2018
}
\description{
Depression Signaling Experiment 2018
}
\details{
DETAILS
}
\keyword{datasets}
